activity/requirements.txt
Django==1.8
Pillow==2.9.0
autopep8==1.1.1
django-contrib-comments==1.6.1
django-extensions==1.5.5
djangorestframework==3.1.3
pep8==1.6.2
six==1.9.0
